Update 1.2 - Temperature Converter
Have added into the program both Celsius and Fahrenheit conversions. Looking to cast Celsius as an integer, however Inteliji does not like the explicit casting of this variable. 
Next update will be looking to simplify the code and take user information at the beginning of the calculations to see if they wish for a specific conversion, rather than providing both conversion at the same time.