Update - 0.3
Added in if statements to make the users journey more organic. "else" statement however on Inteliji is creating an error message which i am unable to resolve. My knowledge of coding does not provide me with enough information to organically resolve this issue, so i shall be looking to resolve this externally before creating another update.